Move leader election config -> pkg/leaderelection
